Modify the index.js  file
